fixing bug : delete Modal in status can't transfer

+++ Kradan-Kanban_Front-End/src/components/Status/DeleteStatus.vue
@@ -1,6 +1,6 @@
 <script setup>
-import { getAllStatus, deleteStatus, transferStatus } from "@/lib/fetchUtils";
-import { ref } from "vue";
+import {getAllStatus, deleteStatus, transferStatus, getAllTasks, getStatusById} from "@/lib/fetchUtils";
+import {onMounted, ref} from "vue";
 
 const emit = defineEmits(["closeModal"]);
 const props = defineProps({
@@ -20,6 +20,14 @@ const statusList = ref(null);
 const error = ref(null);
 const newId = ref(null);
 
+onMounted(async () => {
+  try {
+    statusList.value = await getAllStatus();
+    console.table(statusList.value);
+  } catch (error) {
+  }
+})
+
 // watch(() => props.deleteId, { immediate: true });
 
 // onMounted(() => {
@@ -29,20 +37,32 @@ const newId = ref(null);
 const deleteThisStatus = async () => {
   loading.value = true;
   let res;
-  try {
-    // Check if the status is already in use
-    const inUse = await isStatusInUse(props.deleteId);
-    if (inUse) {
-      // If status is in use, switch to transferring status
-      transfer.value = true;
-      return;
+
+  const inUse = await isStatusInUse(props.deleteId);
+  console.log(await inUse)
+  if (inUse) {
+    // If status is in use, switch to transferring status
+    transfer.value = true;
+    loading.value = false;
+    return true;
+  }
+  if (!inUse) {
+    try {
+      // Check if the status is already in use
+      const inUse = await isStatusInUse(props.deleteId);
+      console.log(await inUse)
+      if (inUse) {
+        // If status is in use, switch to transferring status
+        transfer.value = true;
+        return;
+      }
+      // If not in use, proceed with deletion
+      res = await deleteStatus(props.deleteId);
+    } catch (error) {
+      console.log(error);
+    } finally {
+      emit("closeModal", res);
     }
-    // If not in use, proceed with deletion
-    res = await deleteStatus(props.deleteId);
-  } catch (error) {
-    console.log(error);
-  } finally {
-    emit("closeModal", res);
   }
 };
 
@@ -50,9 +70,12 @@ const closeThisModal = () => {
   emit("closeModal", null);
 };
 
-const isStatusInUse = async (statusId) => {
-  statusList = await getAllStatus();
-  return statusList.some((status) => status.id === statusId && status.inUse);
+const isStatusInUse = async (StatusId) => {
+  let Status = await getStatusById(StatusId)
+  let TaskList = await getAllTasks()
+  return await TaskList.some(task => task.status === Status.name);
+  // statusList.value = await getAllStatus();
+  // return statusList.some((status) => status.id === statusId && status.inUse);
 };
 
 const transferTheStatus = async (newId) => {
@@ -72,26 +95,26 @@ const transferTheStatus = async (newId) => {
     <!-- Delete Modal -->
     <div v-if="!transfer">
       <h1 class="m-2 pb-4 text-2xl font-bold">Delete a Satus</h1>
-      <hr />
+      <hr/>
       <h1 class="itbkk-message font-semibold text-xl p-8">
         Do you want to delete the "{{ deleteTitle }}" Status?
       </h1>
-      <hr />
+      <hr/>
       <div class="flex flex-row-reverse gap-4 mt-5">
         <button
-          @click="closeThisModal"
-          class="itbkk-button-cancel btn btn-outline btn-error basis-1/6"
+            @click="closeThisModal"
+            class="itbkk-button-cancel btn btn-outline btn-error basis-1/6"
         >
           Close
         </button>
         <button
-          @click="deleteThisStatus()"
-          class="itbkk-button-confirm btn btn-outline btn-success basis-1/6"
+            @click="deleteThisStatus()"
+            class="itbkk-button-confirm btn btn-outline btn-success basis-1/6"
         >
           {{ loading ? "" : "Confirm" }}
           <span
-            class="loading loading-spinner text-success"
-            v-if="loading"
+              class="loading loading-spinner text-success"
+              v-if="loading"
           ></span>
         </button>
       </div>
@@ -99,7 +122,7 @@ const transferTheStatus = async (newId) => {
     <!-- Tranfers Modal -->
     <div v-if="transfer">
       <h1 class="m-2 pb-4 text-2xl font-bold">Transfer a Satus</h1>
-      <hr />
+      <hr/>
       <h1 class="itbkk-message font-semibold text-xl p-8">
         There is some task associated with the "{{ deleteTitle }}" Status?
         Tranfer to
@@ -109,31 +132,32 @@ const transferTheStatus = async (newId) => {
           <span class="label-text">Status</span>
         </div>
         <select
-          v-model="newId"
-          class="itbkk-status select select-bordered bg-white"
+            v-model="newId"
+            class="itbkk-status select select-bordered bg-white"
         >
+
           <!-- <option value="No Status" selected>No Status</option> -->
-          <option v-for="status in statusList" :key="status.id" :value="status">
+          <option v-for="status in statusList" :key="status.id" :value="status.id">
             {{ status.name }}
           </option>
         </select>
       </label>
-      <hr />
+      <hr/>
       <div class="flex flex-row-reverse gap-4 mt-5">
         <button
-          @click="closeThisModal"
-          class="itbkk-button-cancel btn btn-outline btn-error basis-1/6"
+            @click="closeThisModal"
+            class="itbkk-button-cancel btn btn-outline btn-error basis-1/6"
         >
           Close
         </button>
         <button
-          @click="transferTheStatus(newId)"
-          class="itbkk-button-confirm btn btn-outline btn-success basis-1/6"
+            @click="transferTheStatus(newId)"
+            class="itbkk-button-confirm btn btn-outline btn-success basis-1/6"
         >
           {{ loading ? "" : "Tranfer" }}
           <span
-            class="loading loading-spinner text-success"
-            v-if="loading"
+              class="loading loading-spinner text-success"
+              v-if="loading"
           ></span>
         </button>
       </div>
